+++ runtime-kotlin/runtime-ports/src/main/kotlin/skillbill/ports/featurespec/FeatureSpecPathResolverPort.kt
@@ -1,0 +1,8 @@
+package skillbill.ports.featurespec
+
+import skillbill.ports.featurespec.model.FeatureSpecPathResolveInput
+import skillbill.ports.featurespec.model.FeatureSpecPathResolveResult
+
+fun interface FeatureSpecPathResolverPort {
+  fun resolve(input: FeatureSpecPathResolveInput): FeatureSpecPathResolveResult
+}